refactor: split commands into separate files
Issue: BTC-1351
