+++ modules/utxo-bin/src/TxParser.ts
@@ -1,5 +1,6 @@
 import { OutputSpend, TransactionStatus } from '@bitgo/blockapis';
 import * as utxolib from '@bitgo/utxo-lib';
+
 import { Parser, ParserNode } from './Parser';
 import { formatSat } from './format';
 import { getParserTxProperties, ParserTx, ParserTxInput, ParserTxOutput } from './ParserTx';
