--- modules/utxo-bin/src/commands.ts
@@ -1,340 +1,0 @@
-/* eslint-disable no-console */
-import * as yargs from 'yargs';
-import * as utxolib from '@bitgo/utxo-lib';
-
-import { Parser, ParserNode } from './Parser';
-import { formatTree } from './format';
-import {
-  fetchOutputSpends,
-  fetchPrevOutputs,
-  fetchPrevOutputSpends,
-  fetchTransactionHex,
-  fetchTransactionStatus,
-  getClient,
-} from './fetch';
-import { TxParser, TxParserArgs } from './TxParser';
-import { AddressParser } from './AddressParser';
-import { parseUnknown } from './parseUnknown';
-import { getParserTxProperties } from './ParserTx';
-import { ScriptParser } from './ScriptParser';
-import {
-  argToString,
-  stringToBuffer,
-  readStringOptions,
-  ReadStringOptions,
-  KeyOptions,
-  keyOptions,
-  getNetworkOptions,
-  getNetworkOptionsDemand,
-} from './args';
-import {
-  formatAddressTree,
-  formatAddressWithFormatString,
-  generateAddress,
-  getAddressPlaceholderDescription,
-  getRange,
-  parseIndexRange,
-} from './generateAddress';
-import { parseXpub } from './bip32';
-
-type OutputFormat = 'tree' | 'json';
-
-type ArgsParseTransaction = ReadStringOptions & {
-  network: utxolib.Network;
-  txid?: string;
-  blockHeight?: number;
-  txIndex?: number;
-  all: boolean;
-  cache: boolean;
-  format: OutputFormat;
-  fetchAll: boolean;
-  fetchStatus: boolean;
-  fetchInputs: boolean;
-  fetchSpends: boolean;
-  finalize: boolean;
-  parseSignatureData: boolean;
-  parseAsUnknown: boolean;
-  parseError: 'throw' | 'continue';
-} & Omit<TxParserArgs, 'parseSignatureData'>;
-
-type ArgsParseAddress = {
-  network?: utxolib.Network;
-  all: boolean;
-  format: OutputFormat;
-  convert: boolean;
-  address: string;
-};
-
-type ArgsParseScript = {
-  network?: utxolib.Network;
-  format: OutputFormat;
-  all: boolean;
-  script: string;
-};
-
-export type ArgsGenerateAddress = KeyOptions & {
-  network: utxolib.Network;
-  chain?: number[];
-  format: string;
-  index?: string[];
-  limit?: number;
-};
-
-type FormatStringArgs = {
-  format: OutputFormat;
-  all: boolean;
-};
-
-function formatString(parsed: ParserNode, argv: yargs.Arguments<FormatStringArgs>): string {
-  switch (argv.format) {
-    case 'json':
-      return JSON.stringify(parsed, null, 2);
-    case 'tree':
-      return formatTree(parsed, { hide: argv.all ? [] : undefined });
-  }
-  throw new Error(`invalid format ${argv.format}`);
-}
-
-export function getTxParser(argv: yargs.Arguments<ArgsParseTransaction>): TxParser {
-  if (argv.all) {
-    return new TxParser({ ...argv, ...TxParser.PARSE_ALL });
-  }
-  return new TxParser({
-    ...argv,
-    parseSignatureData: {
-      script: argv.parseSignatureData,
-      ecdsa: argv.parseSignatureData,
-      schnorr: argv.parseSignatureData,
-    },
-  });
-}
-
-export function getAddressParser(argv: ArgsParseAddress): AddressParser {
-  return new AddressParser(argv);
-}
-
-export function getScriptParser(argv: ArgsParseScript): ScriptParser {
-  return new ScriptParser(argv);
-}
-
-export const cmdParseTx = {
-  command: 'parseTx [path]',
-  aliases: ['parse', 'tx'],
-  describe:
-    'Display transaction components in human-readable form. ' +
-    'Supported formats are Partially Signed Bitcoin Transaction (PSBT), ' +
-    'bitcoinjs-lib encoding (Legacy) or fully signed transaction. ' +
-    'Bytes must be encoded in hex or base64 format.',
-
-  builder(b: yargs.Argv<unknown>): yargs.Argv<ArgsParseTransaction> {
-    return b
-      .options(readStringOptions)
-      .options(getNetworkOptionsDemand())
-      .option('txid', { type: 'string' })
-      .option('blockHeight', { type: 'number' })
-      .option('txIndex', { type: 'number' })
-      .option('fetchAll', { type: 'boolean', default: false })
-      .option('fetchStatus', { type: 'boolean', default: false })
-      .option('fetchInputs', { type: 'boolean', default: false })
-      .option('fetchSpends', { type: 'boolean', default: false })
-      .option('parseScriptAsm', { alias: 'scriptasm', type: 'boolean', default: false })
-      .option('parseScriptData', { alias: 'scriptdata', type: 'boolean', default: false })
-      .option('parseSignatureData', { alias: 'sigdata', type: 'boolean', default: false })
-      .option('parseOutputScript', { type: 'boolean', default: false })
-      .option('parseAsUnknown', {
-        type: 'boolean',
-        default: false,
-        description: 'show plain Javascript object without any post-processing',
-      })
-      .option('maxOutputs', { type: 'number' })
-      .option('vin', { type: 'number' })
-      .array('vin')
-      .option('vout', { type: 'number' })
-      .array('vout')
-      .option('finalize', {
-        type: 'boolean',
-        default: false,
-        description: 'finalize PSBT and parse result instead of PSBT',
-      })
-      .option('all', { type: 'boolean', default: false })
-      .option('cache', {
-        type: 'boolean',
-        default: false,
-        description: 'use local cache for http responses',
-      })
-      .option('format', { choices: ['tree', 'json'], default: 'tree' } as const)
-      .option('parseError', { choices: ['continue', 'throw'], default: 'continue' } as const);
-  },
-
-  async handler(argv: yargs.Arguments<ArgsParseTransaction>): Promise<void> {
-    let data;
-
-    const httpClient = await getClient({ cache: argv.cache });
-
-    if (argv.txid || argv.blockHeight !== undefined || argv.txIndex !== undefined) {
-      data = await fetchTransactionHex(
-        httpClient,
-        {
-          txid: argv.txid,
-          blockHeight: argv.blockHeight,
-          txIndex: argv.txIndex,
-        },
-        argv.network
-      );
-    }
-
-    const string = await argToString(argv, data);
-    if (!string) {
-      throw new Error(`no txdata`);
-    }
-
-    const bytes = stringToBuffer(string, 'hex');
-
-    let tx = utxolib.bitgo.isPsbt(bytes)
-      ? utxolib.bitgo.createPsbtFromBuffer(bytes, argv.network)
-      : utxolib.bitgo.createTransactionFromBuffer(bytes, argv.network, { amountType: 'bigint' });
-
-    const { id: txid } = getParserTxProperties(tx, undefined);
-    if (tx instanceof utxolib.bitgo.UtxoTransaction) {
-      if (argv.txid && txid !== argv.txid) {
-        throw new Error(`computed txid does not match txid argument`);
-      }
-    } else if (argv.finalize) {
-      tx.finalizeAllInputs();
-      tx = tx.extractTransaction();
-    }
-
-    if (argv.parseAsUnknown) {
-      console.log(formatString(parseUnknown(new Parser(), 'tx', tx), argv));
-      return;
-    }
-
-    if (argv.fetchAll) {
-      argv.fetchStatus = true;
-      argv.fetchInputs = true;
-      argv.fetchSpends = true;
-    }
-
-    const parsed = getTxParser(argv).parse(tx, {
-      status: argv.fetchStatus && txid ? await fetchTransactionStatus(httpClient, txid, argv.network) : undefined,
-      prevOutputs: argv.fetchInputs ? await fetchPrevOutputs(httpClient, tx) : undefined,
-      prevOutputSpends: argv.fetchSpends ? await fetchPrevOutputSpends(httpClient, tx) : undefined,
-      outputSpends:
-        argv.fetchSpends && tx instanceof utxolib.bitgo.UtxoTransaction
-          ? await fetchOutputSpends(httpClient, tx)
-          : undefined,
-    });
-
-    console.log(formatString(parsed, argv));
-  },
-} as const;
-
-export const cmdParseAddress = {
-  command: 'parseAddress [address]',
-  aliases: ['address'],
-  describe: 'parse address',
-  builder(b: yargs.Argv<unknown>): yargs.Argv<ArgsParseAddress> {
-    return b
-      .options(getNetworkOptions())
-      .option('format', { choices: ['tree', 'json'], default: 'tree' } as const)
-      .option('convert', { type: 'boolean', default: false })
-      .option('all', { type: 'boolean', default: false })
-      .positional('address', { type: 'string', demandOption: true });
-  },
-
-  handler(argv: yargs.Arguments<ArgsParseAddress>): void {
-    const parsed = getAddressParser(argv).parse(argv.address);
-    console.log(formatString(parsed, argv));
-  },
-} as const;
-
-export const cmdParseScript = {
-  command: 'parseScript [script]',
-  describe: 'parse script',
-  builder(b: yargs.Argv<unknown>): yargs.Argv<ArgsParseScript> {
-    return b
-      .options(getNetworkOptions())
-      .option('format', { choices: ['tree', 'json'], default: 'tree' } as const)
-      .option('all', { type: 'boolean', default: false })
-      .positional('script', { type: 'string', demandOption: true });
-  },
-  handler(argv: yargs.Arguments<ArgsParseScript>): void {
-    const script = stringToBuffer(argv.script, 'hex');
-    const parsed = getScriptParser(argv).parse(script);
-    console.log(formatString(parsed, { ...argv, all: true }));
-  },
-};
-
-export const cmdGenerateAddress = {
-  command: 'generateAddresses',
-  describe: 'generate addresses',
-  builder(b: yargs.Argv<unknown>): yargs.Argv<ArgsGenerateAddress> {
-    return b
-      .options(getNetworkOptionsDemand('bitcoin'))
-      .options(keyOptions)
-      .option('format', {
-        type: 'string',
-        default: '%p0\t%a',
-        description: `Format string. Placeholders: ${getAddressPlaceholderDescription()}`,
-      })
-      .option('chain', { type: 'number', description: 'Address chain' })
-      .array('chain')
-      .option('index', {
-        type: 'string',
-        description: 'Address index. Can be given as a range (e.g. 0-99). Takes precedence over --limit.',
-      })
-      .array('index')
-      .option('limit', {
-        type: 'number',
-        description: 'Alias for --index with range starting at 0 to limit-1.',
-        default: 100,
-      });
-  },
-  handler(argv: yargs.Arguments<ArgsGenerateAddress>): void {
-    let indexRange: number[];
-    if (argv.index) {
-      indexRange = parseIndexRange(argv.index);
-    } else if (argv.limit) {
-      indexRange = getRange(0, argv.limit - 1);
-    } else {
-      throw new Error(`no index or limit`);
-    }
-    for (const address of generateAddress({
-      ...argv,
-      index: indexRange,
-    })) {
-      if (argv.format === 'tree') {
-        console.log(formatAddressTree(address));
-      } else {
-        console.log(formatAddressWithFormatString(address, argv.format));
-      }
-    }
-  },
-};
-
-export const cmdParseXpub = {
-  command: 'parseXpub [xpub]',
-  describe: 'show xpub info',
-  builder(b: yargs.Argv<unknown>): yargs.Argv<
-    {
-      xpub: string;
-      derive?: string;
-    } & FormatStringArgs
-  > {
-    return b
-      .positional('xpub', { type: 'string', demandOption: true })
-      .option('format', { choices: ['tree', 'json'], default: 'tree' } as const)
-      .option('all', { type: 'boolean', default: false })
-      .option('derive', { type: 'string', description: 'show xpub derived with path' });
-  },
-  handler(
-    argv: yargs.Arguments<
-      {
-        xpub: string;
-        derive?: string;
-      } & FormatStringArgs
-    >
-  ): void {
-    console.log(formatString(parseXpub(argv.xpub, { derive: argv.derive }), argv));
-  },
-};
