+++ modules/utxo-bin/src/commands/index.ts
@@ -1,0 +1,5 @@
+export * from './cmdParseTx';
+export * from './cmdParseAddress';
+export * from './cmdParseScript';
+export * from './cmdGenerateAddress';
+export * from './cmdParseXpub';
