+++ modules/utxo-bin/test/bip32.util.ts
@@ -1,5 +1,6 @@
-import * as utxolib from '@bitgo/utxo-lib';
 import * as crypto from 'crypto';
+
+import * as utxolib from '@bitgo/utxo-lib';
 import { BIP32Interface } from 'bip32';
 
 export type Triple<T> = [T, T, T];
